+++ src/main/java/at/hannibal2/skyhanni/SkyHanniMod.java
@@ -29,6 +29,7 @@ import at.hannibal2.skyhanni.features.items.ItemStars;
 import at.hannibal2.skyhanni.features.items.RngMeterInventory;
 import at.hannibal2.skyhanni.features.items.abilitycooldown.ItemAbilityCooldown;
 import at.hannibal2.skyhanni.features.minion.MinionFeatures;
+import at.hannibal2.skyhanni.features.nether.ashfang.AshfangBlazingSouls;
 import at.hannibal2.skyhanni.features.nether.ashfang.AshfangFreezeCooldown;
 import at.hannibal2.skyhanni.features.nether.ashfang.AshfangGravityOrbs;
 import at.hannibal2.skyhanni.features.nether.ashfang.AshfangNextResetCooldown;
@@ -90,6 +91,7 @@ public class SkyHanniMod {
         registerEvent(new AshfangNextResetCooldown());
         registerEvent(new SummoningSoulsName());
         registerEvent(new AshfangGravityOrbs());
+        registerEvent(new AshfangBlazingSouls());
         registerEvent(new ItemStars());
         registerEvent(new MinionFeatures());
         registerEvent(new RealTime());