+++ src/main/java/at/hannibal2/skyhanni/features/nether/ashfang/AshfangBlazingSouls.kt
@@ -16,11 +16,10 @@ import net.minecraftforge.fml.common.eventhandler.SubscribeEvent
 import net.minecraftforge.fml.common.gameevent.TickEvent
 import java.awt.Color
 
-class AshfangGravityOrbs {
+class AshfangBlazingSouls {
 
-    private val texture = "eyJ0ZXh0dXJlcyI6eyJTS0lOIjp7InVybCI6Imh0dHA6Ly90ZXh0dXJlcy5taW5lY3JhZnQubmV" +
-            "0L3RleHR1cmUvMWE2OWNjZjdhZDkwNGM5YTg1MmVhMmZmM2Y1YjRlMjNhZGViZjcyZWQxMmQ1ZjI0Yjc4Y2UyZDQ0YjRhMiJ9fX0="
-    private val orbs = mutableListOf<EntityArmorStand>()
+    private val texture = "eyJ0ZXh0dXJlcyI6eyJTS0lOIjp7InVybCI6Imh0dHA6Ly90ZXh0dXJlcy5taW5lY3JhZnQubmV0L3RleHR1cmUvODI4N2IzOTdkYWY5NTE2YTBiZDc2ZjVmMWI3YmY5Nzk1MTVkZjNkNWQ4MzNlMDYzNWZhNjhiMzdlZTA4MjIxMiJ9fX0="
+    private val souls = mutableListOf<EntityArmorStand>()
 
     @SubscribeEvent
     fun onTick(event: TickEvent.ClientTickEvent) {
@@ -28,37 +27,37 @@ class AshfangGravityOrbs {
 
         Minecraft.getMinecraft().theWorld.loadedEntityList
             .filter { it ->
-                it is EntityArmorStand && it !in orbs && it.inventory
+                it is EntityArmorStand && it !in souls && it.inventory
                     .any { it != null && it.getSkullTexture() == texture }
-            }.forEach { orbs.add(it as EntityArmorStand) }
+            }.forEach { souls.add(it as EntityArmorStand) }
     }
 
     @SubscribeEvent
     fun onRenderWorld(event: RenderWorldLastEvent) {
         if (!isEnabled()) return
 
-        val special = SkyHanniMod.feature.ashfang.gravityOrbsColor
+        val special = SkyHanniMod.feature.ashfang.blazingSoulsColor
 
         val color = Color(SpecialColour.specialToChromaRGB(special), true)
 
         val playerLocation = LocationUtils.playerLocation()
-        for (orb in orbs) {
+        for (orb in souls) {
             if (orb.isDead) continue
             val orbLocation = orb.getLorenzVec()
             event.drawWaypointFilled(orbLocation.add(-0.5, 1.25, -0.5), color)
-            if (orbLocation.distance(playerLocation) < 15) {
+            if (orbLocation.distance(playerLocation) < 10) {
                 //TODO find way to dynamically change color
-                event.drawString(orbLocation.add(0.0, 2.5, 0.0), "§cGravity Orb")
+                event.drawString(orbLocation.add(0.0, 2.5, 0.0), "§bBlazing Soul")
             }
         }
     }
 
     @SubscribeEvent
     fun renderOverlay(event: WorldEvent.Load) {
-        orbs.clear()
+        souls.clear()
     }
 
     private fun isEnabled(): Boolean {
-        return LorenzUtils.inSkyblock && SkyHanniMod.feature.ashfang.gravityOrbs
+        return LorenzUtils.inSkyblock && SkyHanniMod.feature.ashfang.blazingSouls
     }
 }